pyproject.toml
[project]
name = "fatigue-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datetime>=6.0",
    "matplotlib>=3.10.8",
    "opencv-python>=4.13.0.90",
    "pyserial>=3.5",
]
